requirements.txt
locust
redis
psutil
python-dotenv
gTTS
pillow
numpy